Added current date time placeholder function name.
If the function name is matched in UpdateGenerator or InsertGenerator, the value is replaced with the real current datetime.
Using current_datetime as the placeholder.
Updated test xml with an example.
Either valueComputed or valueDate can be used to set the value.

Removed redundant castings in Column.java and unused methods.