Implemented object quoting strategy
Cuurently, only postgres implements a form of quoting whereby mixed case object names and object names that start with numeric characters are quoted.

I added a new property for either a changelog or a changeset.
A user can choose, to keep the existing functionality for backwards compatibility, quote only reserved words or quote all object names.

This allows for a conssitent approach to quoting object names across all databases. I added suuport for seperate setting of the referenced table name and column names in a constraint as well. This allows the individual objects to be quoted if necessary.

Added precondition for checking what quoting strategy is set.
This is used for testing.

Uncommented the liquibase integration tests module in the main pom.
Without this, the integration tests wouldn't run in intellij.