Fixes for running Oracle Integration tests.
Added null checks for object name.
Oracle doesn't support setting a sequence as a default value so removed oracle from that test.

After the Diff refactoring, CreateTableChange is not having its columns serialized.
Issue is due to the method name getColumn in ChangeWithColumns.
This should be getColumns to match the expected getter name for columns.
Without this in CreateTableChange, the columns property won't be considered for serialization
as it won't have a valid read and write method.

I also encountered weird behavior with Oracle whereby an overly long description was correctly
limited to 255 characters but was for some reason being treated as 259 characters by oracle.
This thus resulted in a constraint violation as the length was then too long for the description.
Bit of a hack but I lowered the description length to 250 and that seems to work.