Mark backing index for primary key as handled as well.
In MissingTableChangeGenerator, auto increment primary keys are handled.
The backing index for the primary key should be handled as well.
Use AUTO_INCREMENT as the auto increment clause for H2.
The IDENTITY clause does not allow a constraint name to be specified when used for creating a table.
Removed redundant code from PostgreSQLIntegrationTest.