CORE-1721 HSQL doesn't respect the QUOTE_ALL_OBJECTS flag.
Just removed the HsqlDatabase.escapeObjectName(...) override, as the
superclass' implementation is more complete. Aside from fixing this bug, the
only functional change should be that object names that match DB-reserved words
are no longer always uppercased, which looks like it was a bug, too (though one
hidden by this one).